requirements.txt
opencv-contrib-python
paddlepaddle
paddleocr
pyodbc
Pillow
python-Levenshtein